+++ business/src/main/kotlin/dev/jombi/template/business/reply/dto/RankedChoiceDto.kt
@@ -2,5 +2,5 @@ package dev.jombi.template.business.reply.dto
 
 data class RankedChoiceDto(
     val rank: Int,
-    val choice: Long
+    val choice: Int,
 )
