+++ core/src/main/kotlin/dev/jombi/template/core/reply/entity/RankedChoice.kt
@@ -2,5 +2,5 @@ package dev.jombi.template.core.reply.entity
 
 data class RankedChoice(
     val rank: Int,
-    val choice: Long
+    val choice: Int,
 ) { companion object }
